+++ irs-registry-client/src/main/java/org/eclipse/tractusx/irs/registryclient/central/DigitalTwinRegistryClient.java
@@ -26,6 +26,7 @@ import java.util.List;
 
 import org.eclipse.tractusx.irs.component.assetadministrationshell.AssetAdministrationShellDescriptor;
 import org.eclipse.tractusx.irs.component.assetadministrationshell.IdentifierKeyValuePair;
+import org.eclipse.tractusx.irs.registryclient.decentral.LookupShellsResponse;
 
 /**
  * Digital Twin Registry Rest Client
@@ -45,7 +46,7 @@ public interface DigitalTwinRegistryClient {
      * @param assetIds The key-value-pair of an Asset identifier
      * @return urn uuid string list
      */
-    List<String> getAllAssetAdministrationShellIdsByAssetLink(List<IdentifierKeyValuePair> assetIds);
+    LookupShellsResponse getAllAssetAdministrationShellIdsByAssetLink(List<IdentifierKeyValuePair> assetIds);
 
 }